requirements.txt
finnhub_python==2.4.28
langchain==1.2.15
langchain_core==1.3.0
langchain_openai==1.1.16
langgraph==1.1.9
newsdataapi==0.1.30
python-dotenv==1.2.2
Requests==2.33.1
streamlit==1.56.0
yfinance==1.3.0
